project/02-backend/requirements.txt
flask
flask-cors
pymongo
sentence_transformers
python-dotenv
numpy
google-generativeai